% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_tweet_user.R
\name{code_tweet_user}
\alias{code_tweet_user}
\title{coding and open Twitter user in browser}
\usage{
code_tweet_user(user_df, startrow = 1)
}
\arguments{
\item{user_df}{Data frame from the rest API}

\item{startrow}{define in which row to start. If not specified it starts in the first row}
}
\value{
new data frame with all codes
}
\description{
coding and open user in browser
}
\examples{
\dontrun{
# to open a specific user: input is a df from the search API:
open_tweet_user(user_df, 1)
}

}
