% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_facebook_post.R
\name{open_facebook_post}
\alias{open_facebook_post}
\title{open facebook post in browser}
\usage{
open_facebook_post(post_id, user_name)
}
\arguments{
\item{post_id, }{Vector with post id as strings}

\item{user_name}{String, must be the username in the url}
}
\value{
none, browser opens automatically
}
\description{
Open a specific post in browser
}
\examples{
\dontrun{
# to open a specific facebook post: input is a vector with post ids (strings) 
# and the username from the url
open_facebook_post(facebook_posts_voice[ ,"id"], "TheVoiceOfGermany")
}

}
