% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_pictures_insta.R
\name{open_pictures_insta}
\alias{open_pictures_insta}
\title{open facebook post in browser}
\usage{
open_pictures_insta(my_df)
}
\arguments{
\item{my_df}{A data frame with the columns "id", "comments", and "likes"}
}
\value{
none, browser opens automatically
}
\description{
Get all comments of a user
}
\examples{
\dontrun{
# to open a specific instagram picture: 
# input is a data frame with the columns 
# "id", "comments", and "likes":
open_pictures_insta(ensoie_df)
 }

}
