% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_tweet.R
\name{open_tweet}
\alias{open_tweet}
\title{open tweets in browser}
\usage{
open_tweet(tweets_df)
}
\arguments{
\item{tweets_df}{Data frame from the stream API}
}
\value{
none, browser opens automatically
}
\description{
open tweets in browser
}
\examples{
\dontrun{
# to open a specific tweet: input is a df from the streamAPI:
open_tweet(tweets_df)
}

}

