% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_tweet_rest.R
\name{open_tweet_rest}
\alias{open_tweet_rest}
\title{open tweets in browser from REST API}
\usage{
open_tweet_rest(tweets_df)
}
\arguments{
\item{tweets_df}{Data frame from the REST API}
}
\value{
none, browser opens automatically
}
\description{
open tweets in browser
}
\examples{
\dontrun{
# to open a specific tweet: input is a df from the REST API:
open_tweet_rest(tweets_df)
}

}

