% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_tweet_user.R
\name{open_tweet_user}
\alias{open_tweet_user}
\title{open Twitter user in browser}
\usage{
open_tweet_user(tweets_df)
}
\arguments{
\item{tweets_df}{Data frame from the stream API}
}
\value{
none, browser opens automatically
}
\description{
open user in browser
}
\examples{
\dontrun{
# to open a specific user: input is a df from the streamAPI:
open_tweet_user(tweets_df)
}

}

