% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readkey.R
\name{readkey}
\alias{readkey}
\title{respond to user}
\usage{
readkey()
}
\value{
console command
}
\description{
Stop process, respond to user
}
\examples{
\dontrun{
# responds to user and waits for command
readkey()
}

}
